Update install scripts

+++ scripts/post-install.sh
@@ -1,7 +1,27 @@
+# Test for proper install directory
+if test -d ~/repos/setup.environment; then
+  echo "You cloned this repo to the right directory, great job!"
+else
+  echo "The install script is only made to work when this repo is cloned into the ~/repos directory"
+  exit 1
+fi
+
+# Test for backup directory
+if test -d ~/repos/setup.environment/backups; then
+  echo "Backup directory exists, continuing"
+else
+  echo "Making the backups directory"
+  mkdir ~/repos/setup.environment/backups
+fi
 
 # Destroy current symlinks.
-for file in ~/{.gitconfig,.profile,.bash_profile,.bashrc,.eslintrc,.gitignore_global,composer.json}; do
-  test -h $file && rm $file || echo "$file is not symlink";
+for file in {.gitconfig,.profile,.bash_profile,.bashrc,.eslintrc,.gitignore_global,composer.json}; do
+  if test -h ~/$file; then
+    rm ~/$file
+  else
+    echo "$file is not symlink moving to ~/repos/setup.environment/backups directory"
+    mv ~/$file ~/repos/setup.environment/backups/$file
+  fi
 done
 unset file
 
@@ -15,4 +35,11 @@ ln -s ~/repos/setup.environment/symlinks/.profile ~/.profile
 ln -s ~/repos/setup.environment/symlinks/composer.json ~/composer.json
 
 # Symbolic links for commands
-ln -s /Applications/SourceTree.app/Contents/Resources/stree ~/bin/
+if test -d ~/bin/; then
+  test -h ~/bin/stree && echo "stree cmd exists" || ln -s /Applications/SourceTree.app/Contents/Resources/stree ~/bin/;
+else
+  mkdir ~/bin
+  test -h ~/bin/stree && echo "stree cmd exists" || ln -s /Applications/SourceTree.app/Contents/Resources/stree ~/bin/;
+fi
+
+echo "Completed install"
